fix: move unit test to rule package
Becouse util is a package should not dependent on audit.
